chore(cleanup): harden quality gates default strictness to enforce
