+++ .adr
@@ -1,0 +1,3 @@
+[adr]
+records-dir = docs/decisions
+
