+++ docs/adr-guide.md
@@ -6,12 +6,12 @@
 
 ## Why ADRs?
 
-Architecture Decision Records capture the **context, decision, and consequences** of significant architectural choices. In ADE, ADRs are mandatory for decisions governed by Π.3.1 (Traceable Rationale).
+Architecture Decision Records capture the **context, decision, and consequences** of significant architectural choices. In ADE, ADRs are mandatory for decisions governed by P.3.1 (Traceable Rationale).
 
 ADRs provide:
-- **Institutional memory** — New team members understand past decisions
-- **Axiom traceability** — Decisions link to governing principles
-- **Change management** — Reversing decisions requires new ADRs
+- **Institutional memory** - New team members understand past decisions
+- **Axiom traceability** - Decisions link to governing principles
+- **Change management** - Reversing decisions requires new ADRs
 
 ---
 
@@ -19,15 +19,15 @@ ADRs provide:
 
 | Scenario                               | ADR Required? |
 | -------------------------------------- | ------------- |
-| Adding a new service or component      | ✅ Yes         |
-| Choosing a framework or library        | ✅ Yes         |
-| Changing database schema significantly | ✅ Yes         |
-| Breaking API changes                   | ✅ Yes         |
-| Security-sensitive changes             | ✅ Yes         |
-| Bug fixes                              | ❌ No          |
-| Refactoring (no behavior change)       | ❌ No          |
-| Adding tests                           | ❌ No          |
-| Routine dependency updates             | ❌ No          |
+| Adding a new service or component      | Yes           |
+| Choosing a framework or library        | Yes           |
+| Changing database schema significantly | Yes           |
+| Breaking API changes                   | Yes           |
+| Security-sensitive changes             | Yes           |
+| Bug fixes                              | No            |
+| Refactoring (no behavior change)       | No            |
+| Adding tests                           | No            |
+| Routine dependency updates             | No            |
 
 ---
 
@@ -35,9 +35,9 @@ ADRs provide:
 
 ```
 docs/
-└── adr/
-    ├── 0001-record-architecture-decisions.md
-    ├── 0002-adopt-spec-kit-workflow.md
+└── decisions/
+    ├── 0001-adopt-axiom-driven-engineering.md
+    ├── 0002-test-heuristic-enhancement-and-test-coverage.md
     └── ...
 ```
 
@@ -54,7 +54,7 @@ docs/
 
 ## Governing Postulate
 
-`Π.{X}.{Y}` — {postulate statement}
+`P.{X}.{Y}` - {postulate statement}
 
 ## Context
 
@@ -79,8 +79,8 @@ docs/
 
 ## Axiom Traceability
 
-- **Root Axiom**: `Σ.{X}` — {axiom statement}
-- **Derivation Chain**: `Π.{X}.1` → `Π.{X}.2` → `Π.{X}.{Y}`
+- **Root Axiom**: `S.{X}` - {axiom statement}
+- **Derivation Chain**: `P.{X}.1` -> `P.{X}.2` -> `P.{X}.{Y}`
 - **Constitutional Reference**: Article {N}, §{M}
 
 ## References
@@ -127,7 +127,7 @@ Accepted
 
 ## Governing Postulate
 
-`Π.1.3` — Spec-Kit CLI workflow for specification creation
+`P.1.3` - Spec-Kit CLI workflow for specification creation
 
 ## Context
 
@@ -143,7 +143,7 @@ enforce our specification-first principle.
 ## Decision
 
 We will adopt the Spec-Kit CLI workflow for all feature development. 
-All repositories will include `.specify/` folders for feature specifications.
+All repositories will include .specify/ folders for feature specifications.
 
 ## Consequences
 
@@ -161,8 +161,8 @@ All repositories will include `.specify/` folders for feature specifications.
 
 ## Axiom Traceability
 
-- **Root Axiom**: `Σ.1` — Specification Primacy
-- **Derivation Chain**: `Π.1.1` → `Π.1.2` → `Π.1.3`
+- **Root Axiom**: `S.1` - Specification Primacy
+- **Derivation Chain**: `P.1.1` -> `P.1.2` -> `P.1.3`
 - **Constitutional Reference**: Article I, §1.3
 
 ## References
@@ -175,12 +175,12 @@ All repositories will include `.specify/` folders for feature specifications.
 
 ## Best Practices
 
-1. **Write ADRs at decision time** — Don't wait until after implementation
-2. **Keep context focused** — Explain the problem, not the history of the universe
-3. **Be specific in decisions** — "Use PostgreSQL" not "Use a relational database"
-4. **Acknowledge tradeoffs** — Every decision has consequences
-5. **Link to evidence** — Reference specs, benchmarks, or research
-6. **Include the governing postulate** — This is what makes it ADE
+1. **Write ADRs at decision time** - Don't wait until after implementation
+2. **Keep context focused** - Explain the problem, not the history of the universe
+3. **Be specific in decisions** - "Use PostgreSQL" not "Use a relational database"
+4. **Acknowledge tradeoffs** - Every decision has consequences
+5. **Link to evidence** - Reference specs, benchmarks, or research
+6. **Include the governing postulate** - This is what makes it ADE
 
 ---
