+++ docs/decisions/0001-adopt-axiom-driven-engineering.md
@@ -1,14 +1,12 @@
-# ADR-0001: Adopt Axiom Driven Engineering
+# Adopt Axiom Driven Engineering
 
-## Status
+* Status: accepted
+* Deciders: First-ADE Founding Team
+* Date: 2026-02-06
 
-Accepted
+Technical Story: Foundational methodology setup
 
-## Governing Postulate
-
-`Π.3.1` — ADRs required for all architectural changes
-
-## Context
+## Context and Problem Statement
 
 The First-ADE organization needed a foundational decision-making framework that:
 
@@ -19,66 +17,45 @@ The First-ADE organization needed a foundational decision-making framework that:
 
 Without a clear methodology, engineering decisions become ad-hoc, rationale is lost, and technical debt accumulates invisibly.
 
-### Alternatives Considered
+## Considered Options
 
-1. **Traditional Agile** — Lacks the formal traceability ADE requires
-2. **Pure TDD** — Covers verification but not specification or rationale
-3. **DDD alone** — Strong domain modeling but no test-first mandate
-4. **Custom ad-hoc process** — Inconsistent and undocumented
+1. Traditional Agile - Lacks the formal traceability ADE requires
+2. Pure TDD - Covers verification but not specification or rationale
+3. DDD alone - Strong domain modeling but no test-first mandate
+4. Custom ad-hoc process - Inconsistent and undocumented
 
-## Decision
+## Decision Outcome
 
-We adopt **Axiom Driven Engineering (ADE)** as the foundational methodology for all First-ADE projects.
+Chosen option: We adopt Axiom Driven Engineering (ADE) as the foundational methodology for all First-ADE projects.
 
 ADE consists of:
-- **5 Core Axioms** (Σ.1–Σ.5) — Fundamental truths
-- **3 Orders of Postulates** (Π.X.Y) — Derived practices
-- **Constitutional Governance** — Quality gates and enforcement
-- **ADR Mandates** — Traceability for all decisions
-- **Spec-Kit Lifecycle** — Executable methodology
+* 5 Core Axioms (S.1-S.5) - Fundamental truths
+* 3 Orders of Postulates (P.X.Y) - Derived practices
+* Constitutional Governance - Quality gates and enforcement
+* ADR Mandates - Traceability for all decisions
+* Spec-Kit Lifecycle - Executable methodology
 
 All projects under First-ADE will:
-- Include `.specify/` folders for specifications
-- Follow Red-Green-Refactor commit patterns
-- Maintain ADRs for significant decisions
-- Provide AI context files for agent collaboration
+* Include .specify/ folders for specifications
+* Follow Red-Green-Refactor commit patterns
+* Maintain ADRs for significant decisions
+* Provide AI context files for agent collaboration
 
-## Consequences
+### Positive Consequences
 
-### Positive
-- All decisions are traceable to first principles
-- New contributors understand "why" not just "what"
-- AI agents have explicit guidance via constitutions
-- Technical debt is visible and justified
+* All decisions are traceable to first principles
+* New contributors understand why not just what
+* AI agents have explicit guidance via constitutions
+* Technical debt is visible and justified
 
-### Negative
-- Higher upfront documentation effort
-- Learning curve for ADE concepts
-- May feel bureaucratic for small changes
+### Negative Consequences
 
-### Neutral
-- Shifts responsibility from intuition to explicit reasoning
-- Creates a paper trail of decisions
+* Higher upfront documentation effort
+* Learning curve for ADE concepts
+* May feel bureaucratic for small changes
 
 ## Axiom Traceability
 
-- **Root Axiom**: `Σ.3` — Traceable Rationale
-- **Supporting Axioms**: 
-  - `Σ.1` (Specification Primacy)
-  - `Σ.2` (Deterministic Verification)
-  - `Σ.5` (AI Symbiosis)
-- **Derivation**: First ADR, establishes the framework itself
-- **Constitutional Reference**: This ADR establishes the Constitution
-
-## References
-
-- [MANIFESTO.md](../MANIFESTO.md)
-- [AXIOMS.md](../AXIOMS.md)
-- [POSTULATES.md](../POSTULATES.md)
-- [CONSTITUTION.md](../CONSTITUTION.md)
-- [METHODOLOGY.md](../METHODOLOGY.md)
-
----
-
-**Decision Date**: 2026-02-06  
-**Decided By**: First-ADE Founding Team
+* Root Axiom: S.3 (Traceable Rationale)
+* Supporting Axioms: S.1 (Specification Primacy), S.2 (Deterministic Verification), S.5 (AI Symbiosis)
+* Governing Postulate: P.3.1 (ADRs required for all architectural changes)
